requirements.txt
click==7.1.1
entrypoints==0.3
fastapi==0.54.1
fastapi-utils==0.2.1
flake8==3.7.9
future==0.18.2
h11==0.9.0
httptools==0.1.1
mccabe==0.6.1
numpy==1.18.3
pip-autoremove==0.9.1
pycodestyle==2.5.0
pydantic==1.5.1
pyflakes==2.1.1
pyglet==1.3.2
python-engineio==3.12.1
python-socketio==4.5.1
six==1.14.0
SQLAlchemy==1.3.16
starlette==0.13.2
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
